Added API Gateway and associated Lambda to redirect to Hubspot

+++ deployment/setup.sh
@@ -17,6 +17,7 @@ set -e
 
 ftp_username=$1
 ftp_password=$2
+client_id=$3
 
 echo "Packaging Applications:"
 
@@ -26,17 +27,21 @@ cp -a ../lambdas/. ./packages
 echo "Downloading latest python packages..."
 for directory in ./packages/*/
 do 
-    python3 -m venv ${directory}/venv
-    source ${directory}/venv/bin/activate
-    pip3 install -r ${directory}/requirements.txt
+    if test -f "${directory}/requirements.txt"; then
+        python3 -m venv ${directory}/venv
+        source ${directory}/venv/bin/activate
+        pip3 install -r ${directory}/requirements.txt
+    fi
 done
 
 echo "Creating ZIP file packages..."
 for directory in ./packages/*/
 do 
-    cp -a ${directory}/venv/lib/python*/site-packages/. ${directory}
-    rm -rf ${directory}/venv
-    rm ${directory}/requirements.txt
+    if [ -d "${directory}/venv" ]; then
+        cp -a ${directory}/venv/lib/python*/site-packages/. ${directory}
+        rm -rf ${directory}/venv
+        rm ${directory}/requirements.txt
+    fi
     zip -r ${directory}/function.zip ${directory}
 done
 
@@ -45,7 +50,8 @@ shopt -s extglob
 for directory in ./packages/*/
 do 
     pushd ${directory}
-    rm -v !(*.zip)
+    rm -rv !(*.zip)
+    rm -rf .*
     find -mindepth 1 -maxdepth 1 -type d -exec rm -r {} \
     popd
 done
@@ -59,6 +65,7 @@ echo "Planning Terraform."
 terraform plan \
     -var "ftp_username=${ftp_username}" \
     -var "ftp_password=${ftp_password}" \
+    -var "client_id=${client_id}" \
     -out=output.tfplan
 
 echo "Review Terraform plan then press enter to continue."
